backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
gunicorn==22.0.0
supabase==2.9.1
openai==0.27.8
httpx>=0.26,<0.28
python-dotenv==1.0.1
